# lib/api-zod/src/generated/types/analizOzet.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatGroup } from './statGroup';

export interface AnalizOzet {
  total_mac: number;
  ev_sahibi: StatGroup;
  beraberlik: StatGroup;
  deplasman: StatGroup;
  kg_var: StatGroup;
  ust_25: StatGroup;
  ust_35: StatGroup;
  iy_ust_15: StatGroup;
  iy_ust_05: StatGroup;
  ort_kart: number;
  /** @nullable */
  ort_korner?: number | null;
  ust_10_korner: StatGroup;
  /** @nullable */
  sik_ms?: string | null;
  /** @nullable */
  sik_iy?: string | null;
}