# lib/api-zod/src/generated/types/matchInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MatchInput {
  /** @nullable */
  matchDate?: string | null;
  league: string;
  homeTeam: string;
  awayTeam: string;
  /** @nullable */
  htScore?: string | null;
  ftScore: string;
  /** @nullable */
  previousScore?: string | null;
  /** @nullable */
  yellowCardsHome?: number | null;
  /** @nullable */
  yellowCardsAway?: number | null;
  /** @nullable */
  redCards?: number | null;
  /** @nullable */
  ligSirasiHome?: number | null;
  /** @nullable */
  ligSirasiAway?: number | null;
  /** @nullable */
  ligSirasiTotal?: number | null;
  oddsHome: number;
  oddsDraw: number;
  oddsAway: number;
  /** @nullable */
  altOdds?: number | null;
  /** @nullable */
  ustOdds?: number | null;
  /** @nullable */
  varOdds?: number | null;
  /** @nullable */
  yokOdds?: number | null;
  /** @nullable */
  altOdds35?: number | null;
  /** @nullable */
  ustOdds35?: number | null;
  /** @nullable */
  iyAltOdds15?: number | null;
  /** @nullable */
  iyUstOdds15?: number | null;
  /** @nullable */
  iyAltOdds05?: number | null;
  /** @nullable */
  iyUstOdds05?: number | null;
  /** @nullable */
  avgOddsMin?: number | null;
  /** @nullable */
  avgOddsMax?: number | null;
  /** @nullable */
  imResult?: string | null;
  /** @nullable */
  kornerHome?: number | null;
  /** @nullable */
  kornerAway?: number | null;
}