# lib/api-zod/src/generated/types/similarityQuery.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SimilarityQuery {
  oddsHome: number;
  oddsDraw: number;
  oddsAway: number;
  /** @nullable */
  altOdds?: number | null;
  /** @nullable */
  ustOdds?: number | null;
  /** @nullable */
  varOdds?: number | null;
  /** @nullable */
  yokOdds?: number | null;
  /** @nullable */
  altOdds35?: number | null;
  /** @nullable */
  ustOdds35?: number | null;
  /** @nullable */
  iyAltOdds15?: number | null;
  /** @nullable */
  iyUstOdds15?: number | null;
  /** @nullable */
  iyAltOdds05?: number | null;
  /** @nullable */
  iyUstOdds05?: number | null;
  /** @nullable */
  league?: string | null;
  /** @nullable */
  ligSirasiDiff?: number | null;
  /** @nullable */
  avgCardsTotal?: number | null;
  /** @nullable */
  maxResults?: number | null;
}